borrowings/telegram_utils.py
import os

import requests

from dotenv import load_dotenv


def send_message(text):
    load_dotenv()
    bot_token = os.getenv("TELEGRAM_BOT_TOKEN")
    chat_id = os.getenv("TELEGRAM_CHAT_ID")
    api_url = f"https://api.telegram.org/bot{bot_token}/sendMessage"
    data = {
        "chat_id": chat_id,
        "text": text,
    }
    response = requests.post(api_url, data=data)
    return response
